requirements.txt
Flask==1.0.2
google-api-python-client==1.7.8
google-cloud-datastore==1.8.0
locustio==0.11.0
numpy==1.16.3
pandas==0.24.2
pycountry==18.12.8
requests==2.21.0
tensorflow==1.13.1